requirements/requirements.txt
# SHA1:26a1da68e1a978adfc818c1a75e63642e86609a6
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
appnope==0.1.3
    # via ipython
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
decorator==5.1.1
    # via ipython
executing==1.2.0
    # via stack-data
ipython==8.7.0
    # via -r requirements/requirements.in
jedi==0.18.2
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
numpy==1.24.0
    # via
    #   -r requirements/requirements.in
    #   pandas
    #   pyarrow
pandas==1.5.2
    # via -r requirements/requirements.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.36
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py4j==0.10.9.7
    # via pyspark
pyarrow==10.0.1
    # via -r requirements/requirements.in
pygments==2.13.0
    # via ipython
pyspark==3.5.1
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2022.7
    # via pandas
six==1.16.0
    # via python-dateutil
stack-data==0.6.2
    # via ipython
traitlets==5.8.0
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via prompt-toolkit
delta-spark==3.1.0
kafka-python